refactor: 중복되는 타입 분리
