+++ src/types/CardExpirationDate.ts
@@ -1,0 +1,6 @@
+type CardExpirationDate = {
+  month: string;
+  year: string;
+};
+
+export default CardExpirationDate;
